requirements.txt
aiogram==3.15.0
aiohttp==3.9.1
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
psycopg2-binary==2.9.10